[REF] base: modularize the reporting engines
We currently force the use of WKHTML, since it's hardcoded
we can't add new reporting engines easily.

Here, we:
- modularize the reporting engines
- make accounting use the same generation pipeline
- allow user to choose their reporting engine
  by company/action report/accounting report
[REF] base: modularize the reporting engines.

We currently force the use of WKHTML, since it's hardcoded
we can't add new reporting engines easily.

Here, we:
 - modularize the reporting engines
 - make accounting use the same generation pipeline
 - allow user to choose their reporting engine
   by company/action report/accounting report
