+++ addons/base_report_wkhtmltox/models/__init__.py
@@ -1,0 +1,3 @@
+# Part of Odoo. See LICENSE file for full copyright and licensing details.
+
+from . import ir_actions_report
