+++ addons/marketing_card/__manifest__.py
@@ -3,7 +3,7 @@
     'version': '1.1',
     'category': 'Marketing/Marketing Card',
     'summary': 'Generate dynamic shareable cards',
-    'depends': ['link_tracker', 'mass_mailing', 'website'],
+    'depends': ['link_tracker', 'mass_mailing', 'website', 'base_report_wkhtmltox'],
     'data': [
         'security/marketing_card_groups.xml',
         'security/ir.model.access.csv',
