+++ addons/marketing_card/tests/common.py
@@ -5,7 +5,6 @@ from unittest.mock import patch
 
 from odoo.tests import BaseCase, TransactionCase
 from odoo.tools import BinaryBytes
-from odoo.addons.base.models.ir_actions_report import IrActionsReport
 from odoo.addons.mail.tests.common import mail_new_test_user
 
 
@@ -24,12 +23,12 @@ class MockImageRender(BaseCase):
     def mock_image_renderer(self, collect_params=True):
         self._wkhtmltoimage_bodies = []
 
-        def _ir_actions_report_build_run_wkhtmltoimage(model, bodies, width, height, image_format="jpg"):
+        def _ir_actions_report_build_run_image_engine(engine_name, model, bodies, width, height, image_format="jpg"):
             if collect_params:
                 self._wkhtmltoimage_bodies.extend(bodies)
             return [VALID_JPEG] * len(bodies)
 
-        with patch.object(IrActionsReport, '_run_wkhtmltoimage', _ir_actions_report_build_run_wkhtmltoimage):
+        with patch.object(type(self.env['ir.actions.report']), '_run_image_engine', _ir_actions_report_build_run_image_engine):
             yield
