+++ addons/web/tests/test_reports.py
@@ -4,7 +4,7 @@ from unittest.mock import Mock, patch
 import odoo.tests
 from odoo.exceptions import UserError
 from odoo.http.session import session_store
-from odoo.tests import tagged
+from odoo.tests import no_retry, tagged
 from odoo.tools import mute_logger
 
 from odoo.addons.base.tests.files import PNG_RAW
@@ -92,7 +92,8 @@ class TestReports(odoo.tests.HttpCase):
         self.assertEqual(result.get('record_id'), None, 'wkhtmltopdf must not have been allowed to fetch the image')
         self.assertEqual(result.get('data'), None, 'wkhtmltopdf must not have been allowed to fetch the image')
 
-    @mute_logger('odoo.addons.base.models.ir_actions_report')
+    @mute_logger('odoo.addons.base_report_wkhtmltox.models.ir_actions_report')
+    @no_retry
     def test_report_error_cleanup(self):
         admin = self.env.ref('base.user_admin')
         self.env['ir.ui.view'].create({
