+++ odoo/addons/base/models/ir_actions_report.py
@@ -1,19 +1,11 @@
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
-import functools
 import io
 import json
 import logging
-import os
-import re
-import subprocess
-import tempfile
 import typing
-import unittest
 from ast import literal_eval
 from collections import OrderedDict
-from contextlib import ExitStack, closing
-from itertools import islice
-from urllib.parse import urlparse
+from collections.abc import Sequence
 
 import lxml.html
 import requests
@@ -24,15 +16,12 @@ from PIL import Image, ImageFile
 from odoo import _, api, fields, models, modules, tools
 from odoo.exceptions import AccessError, RedirectWarning, UserError, ValidationError
 from odoo.fields import Domain
-from odoo.http import request
-from odoo.http.session import session_store, update_session_token
-from odoo.tools import config, is_html_empty, parse_version, split_every
+from odoo.tools import is_html_empty
 from odoo.tools.barcode import (
     check_barcode_encoding,
     createBarcodeDrawing,
     get_barcode_font,
 )
-from odoo.tools.misc import find_in_path
 from odoo.tools.pdf import PdfFileReader, PdfFileWriter, PdfReadError
 from odoo.tools.safe_eval import safe_eval, time
 
@@ -40,122 +29,6 @@ from odoo.tools.safe_eval import safe_eval, time
 ImageFile.LOAD_TRUNCATED_IMAGES = True
 
 _logger = logging.getLogger(__name__)
-
-
-def _run_wkhtmltopdf(args):
-    """
-    Runs the given arguments against the wkhtmltopdf binary.
-
-    Returns:
-        The process
-    """
-    bin_path = _wkhtml().bin
-    return subprocess.run(
-        [bin_path, *args],
-        capture_output=True,
-        encoding='utf-8',
-        check=False,
-    )
-
-
-def _split_table(tree, max_rows):
-    """
-    Walks through the etree and splits tables with more than max_rows rows into
-    multiple tables with max_rows rows.
-
-    This function is needed because wkhtmltopdf has a exponential processing
-    time growth when processing tables with many rows. This function is a
-    workaround for this problem.
-
-    :param tree: The etree to process
-    :param max_rows: The maximum number of rows per table
-    """
-    for table in list(tree.iter('table')):
-        prev = table
-        for rows in islice(split_every(max_rows, table), 1, None):
-            sibling = etree.Element('table', attrib=table.attrib)
-            sibling.extend(rows)
-            prev.addnext(sibling)
-            prev = sibling
-
-
-class WkhtmlInfo(typing.NamedTuple):
-    state: typing.Literal['install', 'ok']
-    dpi_zoom_ratio: bool
-    bin: str
-    version: str
-    is_patched_qt: bool
-    wkhtmltoimage_bin: str
-    wkhtmltoimage_version: tuple[str, ...] | None
-
-
-@functools.cache
-def _wkhtml() -> WkhtmlInfo:
-    state = 'install'
-    bin_path = 'wkhtmltopdf'
-    version = ''
-    is_patched_qt = False
-    dpi_zoom_ratio = False
-    try:
-        bin_path = find_in_path('wkhtmltopdf')
-        process = subprocess.Popen(
-            [bin_path, '--version'], stdout=subprocess.PIPE, stderr=subprocess.PIPE
-        )
-    except OSError:
-        _logger.info('You need Wkhtmltopdf to print a pdf version of the reports.')
-    else:
-        _logger.info('Will use the Wkhtmltopdf binary at %s', bin_path)
-        out, _err = process.communicate()
-        version = out.decode('ascii')
-        if '(with patched qt)' in version:
-            is_patched_qt = True
-        match = re.search(r'([0-9.]+)', version)
-        if match:
-            version = match.group(0)
-            if parse_version(version) < parse_version('0.12.0'):
-                _logger.info('Upgrade Wkhtmltopdf to (at least) 0.12.0')
-                state = 'upgrade'
-            else:
-                state = 'ok'
-            if parse_version(version) >= parse_version('0.12.2'):
-                dpi_zoom_ratio = True
-
-            if config['workers'] == 1:
-                _logger.info('You need to start Odoo with at least two workers to print a pdf version of the reports.')
-                state = 'workers'
-        else:
-            _logger.info('Wkhtmltopdf seems to be broken.')
-            state = 'broken'
-
-    wkhtmltoimage_version = None
-    image_bin_path = 'wkhtmltoimage'
-    try:
-        image_bin_path = find_in_path('wkhtmltoimage')
-        process = subprocess.Popen(
-            [image_bin_path, '--version'], stdout=subprocess.PIPE, stderr=subprocess.PIPE
-        )
-    except OSError:
-        _logger.info('You need Wkhtmltoimage to generate images from html.')
-    else:
-        _logger.info('Will use the Wkhtmltoimage binary at %s', image_bin_path)
-        out, _err = process.communicate()
-        match = re.search(rb'([0-9.]+)', out)
-        if match:
-            wkhtmltoimage_version = parse_version(match.group(0).decode('ascii'))
-            if config['workers'] == 1:
-                _logger.info('You need to start Odoo with at least two workers to convert images to html.')
-        else:
-            _logger.info('Wkhtmltoimage seems to be broken.')
-
-    return WkhtmlInfo(
-        state=state,
-        dpi_zoom_ratio=dpi_zoom_ratio,
-        bin=bin_path,
-        version=version,
-        is_patched_qt=is_patched_qt,
-        wkhtmltoimage_bin=image_bin_path,
-        wkhtmltoimage_version=wkhtmltoimage_version,
-    )
 
 
 class IrActionsReport(models.Model):
@@ -178,7 +51,7 @@ class IrActionsReport(models.Model):
     ], required=True, default='qweb-pdf',
     help='The type of the report that will be rendered, each one having its own'
         ' rendering method. HTML means the report will be opened directly in your'
-        ' browser PDF means the report will be rendered using Wkhtmltopdf and'
+        ' browser, PDF means the report will be generated and'
         ' downloaded by the user.')
     report_name = fields.Char(string='Template Name', required=True)
     report_file = fields.Char(string='Report File', required=False, readonly=False, store=True,
@@ -277,17 +150,20 @@ class IrActionsReport(models.Model):
         ], limit=1)
 
     @api.model
-    def get_wkhtmltopdf_state(self):
-        '''Get the current state of wkhtmltopdf: install, ok, upgrade, workers or broken.
-        * install: Starting state.
-        * upgrade: The binary is an older version (< 0.12.0).
-        * ok: A binary was found with a recent version (>= 0.12.0).
-        * workers: Not enough workers found to perform the pdf rendering process (< 2 workers).
-        * broken: A binary was found but not responding.
+    def get_pdf_engine_state(self, engine_name) -> typing.Literal['ok', 'upgrade', 'workers', 'broken', 'install']:
+        """
+        Returns the requested engine status.
+        The state of the pdf engine: install, ok, upgrade, workers or broken.
 
-        :return: wkhtmltopdf_state
-        '''
-        return _wkhtml().state
+        - ok: A binary was found with a recent version.
+        - upgrade: The binary is an older version.
+        - workers: Not enough workers found to perform the pdf rendering process (< 2 workers).
+        - broken: A binary was found but not responding.
+        - install: Starting state.
+
+        :param engine_name: The engine to test
+        """
+        return 'install'
 
     def get_paperformat(self):
         return self.paperformat_id or self.env.company.paperformat_id
@@ -301,88 +177,6 @@ class IrActionsReport(models.Model):
     def _get_report_url(self, layout=None):
         report_url = self.env['ir.config_parameter'].sudo().get_str('report.url')
         return report_url or (layout or self._get_layout() or self).get_base_url()
-
-    @api.model
-    def _build_wkhtmltopdf_args(
-            self,
-            paperformat_id,
-            landscape,
-            specific_paperformat_args=None,
-            set_viewport_size=False):
-        '''Build arguments understandable by wkhtmltopdf bin.
-
-        :param paperformat_id: A report.paperformat record.
-        :param landscape: Force the report orientation to be landscape.
-        :param specific_paperformat_args: A dictionary containing prioritized wkhtmltopdf arguments.
-        :param set_viewport_size: Enable a viewport sized '1024x1280' or '1280x1024' depending of landscape arg.
-        :return: A list of string representing the wkhtmltopdf process command args.
-        '''
-        if landscape is None and specific_paperformat_args and specific_paperformat_args.get('data-report-landscape'):
-            landscape = specific_paperformat_args.get('data-report-landscape')
-
-        command_args = ['--disable-local-file-access']
-        if set_viewport_size:
-            command_args.extend(['--viewport-size', landscape and '1024x1280' or '1280x1024'])
-
-        # Less verbose error messages
-        command_args.extend(['--quiet'])
-
-        # Build paperformat args
-        if paperformat_id:
-            if paperformat_id.format and paperformat_id.format != 'custom':
-                command_args.extend(['--page-size', paperformat_id.format])
-
-            if paperformat_id.page_height and paperformat_id.page_width and paperformat_id.format == 'custom':
-                command_args.extend(['--page-width', str(paperformat_id.page_width) + 'mm'])
-                command_args.extend(['--page-height', str(paperformat_id.page_height) + 'mm'])
-
-            if specific_paperformat_args and 'data-report-margin-top' in specific_paperformat_args:
-                command_args.extend(['--margin-top', str(specific_paperformat_args['data-report-margin-top'])])
-            else:
-                command_args.extend(['--margin-top', str(paperformat_id.margin_top)])
-
-            dpi = None
-            if specific_paperformat_args and specific_paperformat_args.get('data-report-dpi'):
-                dpi = int(specific_paperformat_args['data-report-dpi'])
-            elif paperformat_id.dpi:
-                if os.name == 'nt' and int(paperformat_id.dpi) <= 95:
-                    _logger.info("Generating PDF on Windows platform require DPI >= 96. Using 96 instead.")
-                    dpi = 96
-                else:
-                    dpi = paperformat_id.dpi
-            if dpi:
-                command_args.extend(['--dpi', str(dpi)])
-                if _wkhtml().dpi_zoom_ratio:
-                    command_args.extend(['--zoom', str(96.0 / dpi)])
-
-            if specific_paperformat_args and 'data-report-header-spacing' in specific_paperformat_args:
-                command_args.extend(['--header-spacing', str(specific_paperformat_args['data-report-header-spacing'])])
-            elif paperformat_id.header_spacing:
-                command_args.extend(['--header-spacing', str(paperformat_id.header_spacing)])
-
-            command_args.extend(['--margin-left', str(paperformat_id.margin_left)])
-
-            if specific_paperformat_args and 'data-report-margin-bottom' in specific_paperformat_args:
-                command_args.extend(['--margin-bottom', str(specific_paperformat_args['data-report-margin-bottom'])])
-            else:
-                command_args.extend(['--margin-bottom', str(paperformat_id.margin_bottom)])
-
-            command_args.extend(['--margin-right', str(paperformat_id.margin_right)])
-            if not landscape and paperformat_id.orientation:
-                command_args.extend(['--orientation', str(paperformat_id.orientation)])
-            if paperformat_id.header_line:
-                command_args.extend(['--header-line'])
-            if paperformat_id.disable_shrinking:
-                command_args.extend(['--disable-smart-shrinking'])
-
-        # Add extra time to allow the page to render
-        delay = self.env['ir.config_parameter'].sudo().get_int('report.print_delay') or 1000
-        command_args.extend(['--javascript-delay', str(delay)])
-
-        if landscape:
-            command_args.extend(['--orientation', 'landscape'])
-
-        return command_args
 
     def _prepare_html(self, html, report_model=False):
         '''Divide and recreate the header/footer html by merging all found in html.
@@ -465,180 +259,6 @@ class IrActionsReport(models.Model):
         })
 
         return bodies, res_ids, header, footer, specific_paperformat_args
-
-    def _run_wkhtmltoimage(self, bodies, width, height, image_format="jpg") -> list[bytes | None]:
-        """
-        :param str bodies: valid html documents as strings
-        :param int width: width in pixels
-        :param int height: height in pixels
-        :param image_format: format of the image
-        :type image_format: typing.Literal['jpg', 'png']
-        """
-        if modules.module.current_test:
-            return [None] * len(bodies)
-        wkhtmltoimage_version = _wkhtml().wkhtmltoimage_version
-        if not wkhtmltoimage_version or wkhtmltoimage_version < parse_version('0.12.0'):
-            raise UserError(_('wkhtmltoimage 0.12.0^ is required in order to render images from html'))
-        command_args = [
-            '--disable-local-file-access', '--disable-javascript',
-            '--quiet',
-            '--width', str(width), '--height', str(height),
-            '--format', image_format,
-        ]
-        with ExitStack() as stack:
-            files = []
-            for body in bodies:
-                (input_fd, input_path) = tempfile.mkstemp(suffix='.html', prefix='report_image_html_input.tmp.')
-                (output_fd, output_path) = tempfile.mkstemp(suffix=f'.{image_format}', prefix='report_image_output.tmp.')
-                stack.callback(os.remove, input_path)
-                stack.callback(os.remove, output_path)
-                os.close(output_fd)
-                with closing(os.fdopen(input_fd, 'wb')) as input_file:
-                    input_file.write(body.encode())
-                files.append((input_path, output_path))
-            output_images = []
-            for (input_path, output_path) in files:
-                wkhtmltoimage = [_wkhtml().wkhtmltoimage_bin, *command_args, input_path, output_path]
-                # start and block, no need for parallelism for now
-                completed_process = subprocess.run(wkhtmltoimage, stdout=subprocess.DEVNULL, stderr=subprocess.PIPE, check=False, encoding='utf-8')
-                if completed_process.returncode:
-                    message = _(
-                        'Wkhtmltoimage failed (error code: %(error_code)s). Message: %(error_message_end)s',
-                        error_code=completed_process.returncode,
-                        error_message_end=completed_process.stderr[-1000:],
-                    )
-                    _logger.warning(message)
-                    output_images.append(None)
-                else:
-                    with open(output_path, 'rb') as output_file:
-                        output_images.append(output_file.read())
-        return output_images
-
-    @api.model
-    def _run_wkhtmltopdf(
-            self,
-            bodies,
-            report_ref=False,
-            header=None,
-            footer=None,
-            landscape=False,
-            specific_paperformat_args=None,
-            set_viewport_size=False):
-        '''Execute wkhtmltopdf as a subprocess in order to convert html given in input into a pdf
-        document.
-
-        :param Iterable[str] bodies: The html bodies of the report, one per page.
-        :param report_ref: report reference that is needed to get report paperformat.
-        :param str header: The html header of the report containing all headers.
-        :param str footer: The html footer of the report containing all footers.
-        :param landscape: Force the pdf to be rendered under a landscape format.
-        :param specific_paperformat_args: dict of prioritized paperformat arguments.
-        :param set_viewport_size: Enable a viewport sized '1024x1280' or '1280x1024' depending of landscape arg.
-        :return: Content of the pdf as bytes
-        :rtype: bytes
-        '''
-        paperformat_id = self._get_report(report_ref).get_paperformat() if report_ref else self.get_paperformat()
-
-        # Build the base command args for wkhtmltopdf bin
-        command_args = self._build_wkhtmltopdf_args(
-            paperformat_id,
-            landscape,
-            specific_paperformat_args=specific_paperformat_args,
-            set_viewport_size=set_viewport_size)
-
-        files_command_args = []
-
-        with ExitStack() as stack:
-
-            # Passing the cookie to wkhtmltopdf in order to resolve internal links.
-            if request and request.db:
-                # Create a temporary session which will not create device logs
-                temp_session = session_store().new()
-                temp_session.update({
-                    **request.session,
-                    'debug': '',
-                    '_trace_disable': True,
-                })
-                if temp_session.uid:
-                    update_session_token(temp_session, self.env)
-                session_store().save(temp_session)
-                stack.callback(session_store().delete, temp_session)
-
-                base_url = self._get_report_url()
-                domain = urlparse(base_url).hostname
-                cookie = f'session_id={temp_session.sid}; HttpOnly; domain={domain}; path=/;'
-                cookie_jar_file = stack.enter_context(tempfile.NamedTemporaryFile(suffix='.txt', prefix='report.cookie_jar.tmp.', delete_on_close=False))
-                with closing(cookie_jar_file):
-                    cookie_jar_file.write(cookie.encode())
-                command_args.extend(['--cookie-jar', cookie_jar_file.name])
-
-            if header:
-                head_file = stack.enter_context(tempfile.NamedTemporaryFile(suffix='.html', prefix='report.header.tmp.', delete_on_close=False))
-                with closing(head_file):
-                    head_file.write(header.encode())
-                files_command_args.extend(['--header-html', head_file.name])
-            if footer:
-                foot_file = stack.enter_context(tempfile.NamedTemporaryFile(suffix='.html', prefix='report.footer.tmp.', delete_on_close=False))
-                with closing(foot_file):
-                    foot_file.write(footer.encode())
-                files_command_args.extend(['--footer-html', foot_file.name])
-
-            paths = []
-            body_idx = 0
-            for body_idx, body in enumerate(bodies):
-                prefix = f'report.body.tmp.{body_idx}.'
-                body_file = stack.enter_context(tempfile.NamedTemporaryFile(suffix='.html', prefix=prefix, delete_on_close=False))
-                with closing(body_file):
-                    # HACK: wkhtmltopdf doesn't like big table at all and the
-                    #       processing time become exponential with the number
-                    #       of rows (like 1H for 250k rows).
-                    #
-                    #       So we split the table into multiple tables containing
-                    #       500 rows each. This reduce the processing time to 1min
-                    #       for 250k rows. The number 500 was taken from opw-1689673
-                    if len(body) < 4 * 1024 * 1024:  # 4Mib
-                        body_file.write(body.encode())
-                    else:
-                        tree = lxml.html.fromstring(body)
-                        _split_table(tree, 500)
-                        body_file.write(lxml.html.tostring(tree))
-                paths.append(body_file.name)
-
-            pdf_report = stack.enter_context(tempfile.NamedTemporaryFile(suffix='.pdf', prefix='report.tmp.', delete_on_close=False))
-            pdf_report.close()
-            pdf_report_path = pdf_report.name
-
-            process = _run_wkhtmltopdf(command_args + files_command_args + paths + [pdf_report_path])
-            err = process.stderr
-
-            match process.returncode:
-                case 0:
-                    pass
-                case 1:
-                    if body_idx:
-                        if not _wkhtml().is_patched_qt:
-                            if modules.module.current_test:
-                                raise unittest.SkipTest("Unable to convert multiple documents via wkhtmltopdf using unpatched QT")
-                            raise UserError(_("Tried to convert multiple documents in wkhtmltopdf using unpatched QT"))
-
-                    _logger.warning("wkhtmltopdf: %s", err)
-                case c:
-                    message = _(
-                        'Wkhtmltopdf failed (error code: %(error_code)s). Memory limit too low or maximum file number of subprocess reached. Message : %(message)s',
-                        error_code=c,
-                        message=err[-1000:],
-                    ) if c == -11 else _(
-                        'Wkhtmltopdf failed (error code: %(error_code)s). Message: %(message)s',
-                        error_code=c,
-                        message=err[-1000:],
-                    )
-                    _logger.warning(message)
-                    raise UserError(message)
-
-            with open(pdf_report_path, 'rb') as pdf_document:
-                pdf_content = pdf_document.read()
-
-        return pdf_content
 
     @api.model
     def _get_report_from_name(self, report_name):
@@ -847,19 +467,41 @@ class IrActionsReport(models.Model):
                     'attachment': attachment,
                 }
 
-        # Call 'wkhtmltopdf' to generate the missing streams.
+        # Call the defined PDF engine to generate the missing streams.
         res_ids_wo_stream = [res_id for res_id, stream_data in collected_streams.items() if not stream_data['stream']]
         all_res_ids_wo_stream = res_ids if has_duplicated_ids else res_ids_wo_stream
-        is_whtmltopdf_needed = not res_ids or res_ids_wo_stream
 
-        if is_whtmltopdf_needed:
+        engine_name = self._get_pdf_engine(report_sudo)
+        if engine_name == 'html':
+            # PDF generation is not setup
 
-            if self.get_wkhtmltopdf_state() == 'install':
-                # wkhtmltopdf is not installed
-                # the call should be catched before (cf /report/check_wkhtmltopdf) but
+            # Disable the debug mode in the PDF rendering in order to not split the
+            # assets bundle into separated files to load. This is done because of
+            # an issue in wkhtmltopdf failing to load the CSS/Javascript resources
+            # in time.
+            # Without this, the header/footer of the reports randomly disappear
+            # because the resources files are not loaded in time.
+            # https://github.com/wkhtmltopdf/wkhtmltopdf/issues/2083
+            additional_context = {'debug': False}
+            data.setdefault("debug", False)
+            return self.with_context(**additional_context) \
+                ._render_qweb_html(report_ref, all_res_ids_wo_stream, data=data)[0]
+
+        is_pdf_engine_required = not res_ids or res_ids_wo_stream
+        if is_pdf_engine_required:
+            engine_state = self.get_pdf_engine_state(engine_name)
+            if engine_state == 'install':
+                # pdf engine is not installed
+                # the call should be catched before (cf /report/get_pdf_engine_state) but
                 # if get_pdf is called manually (email template), the check could be
                 # bypassed
-                raise UserError(_("Unable to find Wkhtmltopdf on this system. The PDF can not be created."))
+                raise UserError(_("Unable to find the pdf engine on this system. The PDF can not be created."))
+            elif engine_state == 'broken':
+                # pdf engine is installed but not working
+                raise UserError(_("The pdf engine on this system is not working."))
+            elif engine_state == 'workers':
+                # the pdf engine is installed but not enough workers are available
+                raise UserError(_("Not enough workers to generate the PDF. Please try another engine or upgrade your setup."))
 
             # Disable the debug mode in the PDF rendering in order to not split the assets bundle
             # into separated files to load. This is done because of an issue in wkhtmltopdf
@@ -872,7 +514,7 @@ class IrActionsReport(models.Model):
 
             html = self.with_context(**additional_context)._render_qweb_html(report_ref, all_res_ids_wo_stream, data=data)[0]
 
-            bodies, html_ids, header, footer, specific_paperformat_args = report_sudo.with_context(**additional_context)._prepare_html(html, report_model=report_sudo.model)
+            pdf_content, html_ids = self._run_pdf_engine(engine_name, html, report_ref=report_ref, landscape=self.env.context.get('landscape'))
 
             if not has_duplicated_ids and report_sudo.attachment and set(res_ids_wo_stream) != set(html_ids):
                 raise UserError(_(
@@ -881,18 +523,7 @@ class IrActionsReport(models.Model):
                     " attributes 'data-oe-model' and 'data-oe-id' as part of the div with 'article' classname.",
                     report_sudo.name,
                 ))
-
-            pdf_content = self._run_wkhtmltopdf(
-                bodies,
-                report_ref=report_ref,
-                header=header,
-                footer=footer,
-                landscape=self.env.context.get('landscape'),
-                specific_paperformat_args=specific_paperformat_args,
-                set_viewport_size=self.env.context.get('set_viewport_size'),
-            )
             pdf_content_stream = io.BytesIO(pdf_content)
-
             # Printing a PDF report without any records. The content could be returned directly.
             if has_duplicated_ids or not res_ids:
                 return {
@@ -1134,6 +765,72 @@ class IrActionsReport(models.Model):
             })
         data['is_html_empty'] = is_html_empty
         return data
+
+    def _get_pdf_engine(self, report=None, default_engine='wkhtmltopdf') -> str:
+        """Resolve the PDF engine name from report settings and db's fallback.
+
+        This helper is shared by modules that manually call low-level PDF rendering
+        methods and still use account reports.
+        """
+        report_type = (getattr(report, 'report_type', '') or '').lower()
+        engine_name = (
+            report_type.removeprefix('qweb-pdf-')
+            if report_type.startswith('qweb-pdf-')
+            else self.env['ir.config_parameter'].sudo().get_str('report.pdf_engine_default') or 'html'
+            )  # Low-level PDF rendering methods require an explicit PDF engine.
+        return engine_name if engine_name and engine_name != 'html' else default_engine
+
+    def _run_image_engine(self, engine_name: str, bodies: Sequence[str], width: int, height: int, image_format: str = "jpg") -> list[bytes | None]:
+        """ Render an image using a PDF engine.
+
+        :param engine_name: the name of the image rendering engine to use
+        :param bodies: valid html documents as strings
+        :param width: width in pixels
+        :param height: height in pixels
+        :param image_format: format of the image (usually jpg or png)
+        """
+        e = NotImplementedError("No image rendering engine installed")
+        raise e
+
+    @api.model
+    def _run_pdf_engine(
+        self,
+        engine_name: str,
+        html: str,
+        report_ref: str | typing.Literal[False] = False,
+        landscape: bool = False,
+        **kwargs,
+    ) -> tuple[bytes, list[int]]:
+        """Render the HTML using the associated PDF engine.
+
+        :param engine_name: The name of the pdf engine to use.
+        :param html: The html to convert to pdf.
+        :param report_ref: The report reference (opt.).
+        :param landscape: If the pdf should be in landscape mode (opt.).
+        :param kwargs: Additional arguments to pass to the pdf engine. (e.g. page size, margins, etc.)
+        :return: a tuple of (pdf, html_ids) where pdf is the generated pdf content and html ids to separate the
+            pdf content.
+        """
+        # If we reach here it means no module handled the pdf rendering with the given engine_name.
+        e = NotImplementedError(f"Unknown PDF engine: {engine_name}")
+        raise e
+
+    @api.model
+    def _run_pdf_engine_without_processing(
+        self,
+        engine_name: str,
+        bodies: Sequence[str],
+        report_ref: str | typing.Literal[False] = False,
+        *,
+        header: str | None = None,
+        footer: str | None = None,
+        landscape: bool = False,
+        specific_paperformat_args: dict | None = None,
+        **kwargs,
+    ) -> bytes:
+        # If we reach here it means no module handled the pdf rendering with the given engine_name.
+        e = NotImplementedError(f"Unknown PDF engine: {engine_name}")
+        raise e
 
     @api.model
     def _render(self, report_ref, res_ids, data=None):
